Final updates to integrate db
- Added update session func to db to be able record changes
- Added new method in session to create unique ids

+++ digital-deck-server/services/utilities.js
@@ -1,14 +1,34 @@
 const Session = require('../models/session');
-const sessionData = require('../models/dbConnection');
+const Player = require('../models/player');
+const Table = require('../models/table');
+const {getDB} = require('../models/dbConnection');
 
-
+let concSessions = [];
+let sessionData;
 /**
  * Cleans sessions from the database
  * 
  * @return [array of {json} ] The array of json containing information of all sessions
  */
 function cleanSession(session) {
+    session.players.forEach((player) => Object.assign(new Player(), player));
+    session.table = Object.assign(new Table(), session.table);
     delete session._id;
+}
+
+/**
+ * Updates sessions in the database
+ */
+async function updateSession(session) {
+    const query = {'sessionId': session.sessionId};
+    if (process.env.NODE_ENV !== 'test')
+    {
+        if (sessionData == undefined)
+        {
+            sessionData = await getDB();
+        }
+        sessionData.replaceOne(query, session);
+    }
 }
 
 /**
@@ -17,10 +37,20 @@ function cleanSession(session) {
  * @return [array of {json} ] The array of json containing information of all sessions
  */
 async function getConcSessions()
-{
-    var sessions = await sessionData.find();
-    console.log(sessions);
-    sessions.forEach((session) => {
+{   
+    if (process.env.NODE_ENV == 'test')
+    {
+        return concSessions;
+    }
+    if (sessionData == undefined)
+    {
+        sessionData = await getDB();
+    }
+    var sessions = [];
+    const cursor = await sessionData.find();
+    sessions = await cursor.toArray();
+    sessions.forEach(session => {
+        console.log(session);
         session = cleanSession(session);
     });
     console.log(sessions);
@@ -32,8 +62,20 @@ async function getConcSessions()
  */
 async function clearConcSessions()
 {
-    await sessionData.deleteMany();
-    Session.id = 0;
+    if (process.env.NODE_ENV == 'test')
+    {
+        concSessions = [];
+        Session.id = 0;
+    }
+    else
+    {
+        if (sessionData == undefined)
+        {
+            sessionData = await getDB();
+        }
+        await sessionData.deleteMany();
+        Session.id = 0;
+    }    
 }
 
 /**
@@ -43,7 +85,18 @@ async function clearConcSessions()
  */
 async function addSession(session)
 {
-    await sessionData.insertOne(session);
+    if (process.env.NODE_ENV == 'test')
+    {
+        concSessions.push(session);
+    }
+    else
+    {
+        if (sessionData == undefined)
+        {
+            sessionData = await getDB();
+        }
+        await sessionData.insertOne(session);
+    }
 }
 
 /**
@@ -53,9 +106,23 @@ async function addSession(session)
  */
 async function getSession(id)
 {
-    const query = {sessionId: id};
-    const session = await sessionData.findOne(query);
-    return session;
+    if (process.env.NODE_ENV == 'test')
+    {
+        return concSessions.find(session => session.sessionId === id);
+    }
+    else
+    {
+        if (sessionData == undefined)
+        {
+            sessionData = await getDB();
+        }
+        const query = {sessionId: id};
+        const session = await sessionData.findOne(query);
+        session.players.forEach((player) => Object.assign(new Player(), player));
+        session.table = Object.assign(new Table(), session.table);
+        return session;        
+    }
+
 }
 
 /**
@@ -66,7 +133,15 @@ async function getSession(id)
  */
 function getPlayer(session, id)
 {
-    return session.players.find(player => player.playerId === id);
+    if (process.env.NODE_ENV == 'test')
+    {
+        return session.players.find(player => player.playerId === id);
+    }
+    else
+    {
+        const player = session.players.find(player => player.playerId === id);
+        return Object.assign(new Player(), player);
+    }
 }
 
-module.exports = {getConcSessions, clearConcSessions, addSession, getSession, getPlayer, cleanSession};
+module.exports = {getConcSessions, clearConcSessions, addSession, getSession, getPlayer, cleanSession, updateSession};
